Adapts HarddriveStorageTest and SanStorageTest
- Changes performed after Host storage redesign
- Fix test failures
