+++ cloudsim-plus/src/test/java/org/cloudsimplus/integrationtests/CloudletSchedulerTimeSharedWithMoreCloudletsThanPEs.java
@@ -29,6 +29,7 @@ import org.cloudbus.cloudsim.core.CloudSim;
 import org.cloudbus.cloudsim.schedulers.cloudlet.CloudletSchedulerTimeShared;
 import org.cloudbus.cloudsim.schedulers.vm.VmSchedulerSpaceShared;
 import org.cloudbus.cloudsim.utilizationmodels.UtilizationModel;
+import org.cloudbus.cloudsim.utilizationmodels.UtilizationModelDynamic;
 import org.cloudbus.cloudsim.utilizationmodels.UtilizationModelFull;
 import org.cloudsimplus.builders.BrokerBuilderDecorator;
 import org.cloudsimplus.builders.HostBuilder;
@@ -42,10 +43,12 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 /**
  *
  * An Integration Test (IT) running a simulation scenario with 1 PM of 2 PEs,
- * 1 VMs of 2 PEs and 4 cloudlet in that VM.
+ * 1 VM of 2 PEs and 4 cloudlet in that VM.
  * The VM uses a {@link CloudletSchedulerTimeShared}. As the number of Cloudlets
  * is the double of VM's PEs, all cloudlets will spend the double of the
  * time to finish, because they will concur for CPU.
+ * Since all 4 Cloudlets runs in the same VM,
+ * they are configure to use just 25% of the VM RAM each one.
  *
  * @author Manoel Campos da Silva Filho
  */
@@ -85,6 +88,7 @@ public final class CloudletSchedulerTimeSharedWithMoreCloudletsThanPEs {
         brokerBuilder.getCloudletBuilder()
             .setLength(CLOUDLET_LENGTH)
             .setUtilizationModelCpu(utilizationModel)
+            .setUtilizationModelRam(new UtilizationModelDynamic(0.25))
             .setPEs(CLOUDLET_PES)
             .createAndSubmit(NUMBER_OF_CLOUDLETS);
     }
